pyproject.toml
[project]
name = "graphql_http"
dynamic = ["version"]
description = "HTTP Server for GraphQL."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
authors = [
    {name = "Robert Parker", email = "rob@parob.com"}
]
keywords = ["GraphQL", "HTTP", "Server", "starlette"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "graphql-core",
    "graphql-api",
    "packaging",
    "graphql-schema-diff",
    "pyjwt[crypto]",
    "starlette",
    "uvicorn",
    "python-multipart",
    "httpx",
]

[project.urls]
Homepage = "https://gitlab.com/parob/graphql-http"
Repository = "https://gitlab.com/parob/graphql-http"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "coverage",
    "flake8",
    "flake8-pyproject",
    "autopep8>=2.3.2",
]

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.python.org/simple"

[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["graphql_http"]
include-package-data = true

[tool.setuptools.package-data]
graphql_http = [
    "graphiql/index.html",
    "service/service_manager_default.graphql"
]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.flake8]
max-line-length = 170
ignore = [
    "E203",  # whitespace before ':' (conflicts with black)
    "W503",  # line break before binary operator (conflicts with black)
]
exclude = [".venv"]

